Rename manager to avoid static linking problems
