+++ src/configuration/configuration_manager.cpp
@@ -14,10 +14,10 @@
 #include "configuration/actions_parser.hpp"
 #include "configuration/common/common.hpp"
 #include "configuration/common/configuration.hpp"
+#include "configuration/configuration_manager.hpp"
 #include "configuration/data_parser.hpp"
 #include "configuration/exclusion_parser.hpp"
 #include "configuration/legacy_rule_parser.hpp"
-#include "configuration/manager.hpp"
 #include "configuration/processor_parser.hpp"
 #include "configuration/rule_override_parser.hpp"
 #include "configuration/rule_parser.hpp"
