+++ src/configuration/configuration_manager.hpp
